Refactoring mixin modules
