+++ lib/tumugi/command/show.rb
@@ -1,13 +1,12 @@
 require 'graphviz'
 require 'tmpdir'
 require 'fileutils'
-
-require 'tumugi/helper'
+require 'tumugi/mixin/listable'
 
 module Tumugi
   module Command
     class Show
-      include Tumugi::Helper
+      include Tumugi::Mixin::Listable
 
       @@supported_formats = ['dot', 'png', 'jpg', 'svg', 'pdf']
