+++ lib/tumugi/mixin/listable.rb
@@ -1,0 +1,17 @@
+module Tumugi
+  module Mixin
+    module Listable
+      def list(obj)
+        if obj.nil?
+          []
+        elsif obj.is_a?(Array)
+          obj
+        elsif obj.is_a?(Hash)
+          obj.map { |k,v| v }
+        else
+          [obj]
+        end
+      end
+    end
+  end
+end
