+++ lib/tumugi/mixin/task_helper.rb
@@ -1,0 +1,12 @@
+require 'tumugi/plugin'
+
+module Tumugi
+  module Mixin
+    module TaskHelper
+      def target(type, *args)
+        klass = Tumugi::Plugin.lookup_target(type)
+        klass.new(*args)
+      end
+    end
+  end
+end
