+++ lib/tumugi/task.rb
@@ -1,10 +1,10 @@
-require 'tumugi/helper'
-require 'tumugi/mixin/task_mixin'
+require 'tumugi/mixin/listable'
+require 'tumugi/mixin/task_helper'
 
 module Tumugi
   class Task
-    include Tumugi::Helper
-    include Tumugi::TaskMixin
+    include Tumugi::Mixin::Listable
+    include Tumugi::Mixin::TaskHelper
 
     attr_accessor :state # :pending, :running, :completed, :skipped
